Adicionado validações e mascaras no Cadastro de Serviços.

+++ database/migrations/2021_08_31_223051_create_servicos_table.php
@@ -18,7 +18,7 @@ class CreateServicosTable extends Migration
             $table->string('nome');
             $table->decimal('valor_minimo', 10, 2);
             $table->integer('quantidade_horas');
-            $table->decimal('porcentagem');
+            $table->integer('porcentagem');
             $table->decimal('valor_quarto', 10, 2);
             $table->integer('horas_quarto');
             $table->decimal('valor_sala', 10, 2);
